Create flag ceph backend configured after restore
During the restore procedure the flag
/etc/platform/.node_ceph_configured is missing at the end of the
restore when running the restore playbook with flag wipe_ceph_osd
set to false.

When unlocking the controller-0 after the restore playbook, the puppet
manifest is trying to reconfigure the ceph-mon filesystem for AIO-DX
systems. This is leading to Ceph cluster data corruption.

To fix this, the flag is now set at the end of the restore procedure.

Test Plan:
  PASS: AIO-DX bootstrap
  PASS: AIO-DX B&R with ceph backend configured

Partial-Bug: 2031542

Change-Id: I80651a125b9d6e21efa1d392c6522c136c72d7a7
Signed-off-by: Felipe Sanches Zanoni <Felipe.SanchesZanoni@windriver.com>