Improved system-local-ca check and install
Included support to check multiple ICAs bundled together (verifying
requirements for each one individually) and have override and default
variables for ICA or RCA expiration checks (default is 3 years for
both). Also, include check to verify if the ICA TLS cert/key match
each other.

Changed the behavior to only install the RCA as trusted (before we
installed both the ICA and the RCA). Also made the install optional
so the role can be used to verify the certificates in the first steps
of the bootstrap. With this, the user can fix the provided
certificates if a problem is detected without waiting until the end
of the bootstrap (where the installation of the RCA will be made).

Changed the naming to better reflect the intention of the role.

Test plan:
PASS: Perform cert-manager migration using:
      - RCA / single ICA;
      - RCA / multiple ICAs in files;
      - RCA / multiple ICAs with one with short expiration date
        (should fail with a message regarding the short expiration);
      - RCA / multiple ICAs with one with short expiration date but
        use override flag to allow short expiration date;
      - RCA with short expiration date / single ICA;
        (should fail with a message regarding the short expiration);
      - RCA with short expiration date but use override flag to allow
        short expiration date/ single ICA;

PASS: Deploy AIO-SX providing the CA certs for 'system-local-ca' and
      enable feature flag. ICA file should contain multiple bundled
      ICAs.The CAs provided must be within the expected expiration
      date. Verify:
      - HTTPS is enabled and openstack public endpoints change into it
        after unlocking the controller.
      - The target certificates are issued by 'system-local-ca', and
        are managed by cert-manager;
      - The 'system-local-ca' has the provided certs data.
      - The certificates in /etc/ssl/private are correct.
      - It's possible to log into the local Docker Registry.
      - Horizon is working as expected.

PASS: Deploy AIO-SX w/o providing the CA certs for 'system-local-ca'
      and enable feature flag. Verify:
      - HTTPS is enabled and openstack public endpoints change into it
        after unlocking the controller.
      - The target certificates are issued by 'system-local-ca', and
        are managed by cert-manager;
      - 'system-local-ca' has the Kubernetes Root CA data.
      - The certificates in /etc/ssl/private are correct.
      - It's possible to log into the local Docker Registry.
      - Horizon is working as expected.

PASS: Deploy AIO-SX providing the CA certs for 'system-local-ca' and
      enable feature flag. ICA file should contain multiple bundled
      ICAs. One of the ICAs provided must have a short expiration
      date. Verify installation fail at the beginning of bootstrap,
      under 'bootstrap/validate-config'.

PASS: Deploy AIO-SX providing the CA certs for 'system-local-ca' and
      enable feature flag. ICA file should contain multiple bundled
      ICAs. One of the ICAs provided must have a short expiration
      date. Provide 'ica_duration' override as to make the
      expiration after the fail limit. Verify that the bootstrap is
      successful.

PASS: Deploy AIO-SX providing the CA certs for 'system-local-ca' and
      enable feature flag. ICA file should contain the same RCA cert
      as in the RCA file. Verify bootstrap fail with a message that
      the ICA must have an ICA.

PASS: Deploy AIO-SX providing the CA certs for 'system-local-ca' and
      enable feature flag. ICA key file should contain a key that
      doesn't match the ICA cert provided. Verify bootstrap fail with
      a message that the ICA TLS cert/key pair must match.

Story: 2009811
Task: 48907

Change-Id: I6a835990a6e3606f5d716dd8ed7225cf60d4bbb9
Signed-off-by: Marcelo Loebens <Marcelo.DeCastroLoebens@windriver.com>