Fix: Migrate CentOS home dir during upgrade
A rare issue was detected on a single system.
The tar command that extracts the home dir was erroring with RC=2.

Presumably this is because the absolute path to the home
directory between CentOS and Debian changed: /home -> /var/home.
It was not determined why this only caused an error on a
single system so far.

To fix this, only the contents of the home directory will be extracted
and not the directory itself.

TEST PLAN
PASS: Optimized upgrade on AIO-SX, stx6 to stx8
* On affected
* On unaffected system

Closes-Bug: https://bugs.launchpad.net/starlingx/+bug/2043280
Signed-off-by: Joshua Kraitberg <joshua.kraitberg@windriver.com>
Change-Id: I72886de3d30f33e4d5f6febb0de5080b69f27b23