Do not restore Ceph crush map when wiping Ceph OSD disks
When running the restore playbook, the Ceph crush map was being
restored even if the flag wipe_ceph_osds was set to true.

The restore playbook was not checking the status of the wipe_ceph_osds
flag. A check is added to the whole block. Now the Ceph crush map is
only restored if the flag is set to false and there is Ceph backend
configured.

Ceph will be reconfigured during the unlock process of each node.

Test Plan:
  PASS: B&R AIO-DX with wipe_ceph_osds=false
  PASS: B&R AIO-DX with wipe_ceph_osds=true
  PASS: B&R Standard with wipe_ceph_osds=false
  PASS: B&R Standard with wipe_ceph_osds=true
  PASS: B&R Storage with wipe_ceph_osds=false
  PASS: B&R Storage with wipe_ceph_osds=true

Closes-bug: 2043412

Change-Id: Ib4e4c6933bf7ff6b8f23d994f19a6e79c01dd2b1
Signed-off-by: Felipe Sanches Zanoni <Felipe.SanchesZanoni@windriver.com>