Define default k8s rootca certificate of a subcloud
This commit adds the default k8s rootca certificate of a subcloud if
not be set in the bootstrap values, ensures the kube rootca
certificate can be synchronized between a central cloud and a subcloud
during bootstrapping the subcloud.

Test Plan:
1. Passed - Build new image, bootstrap up an AIOSX.
2. Passed - Deploy distributed cloud with an AIOSX subcloud, verify
the subcloud has the same Kubernetes root CA cert as the central
cloud.
3. Passed - Re-deploy the subcloud, verify the subcloud has the same
Kubernetes root CA cert as the central cloud.

Story: 2010852
Task: 49104

Signed-off-by: Yuxing Jiang <Yuxing.Jiang@windriver.com>
Change-Id: Ie9e783fc44308bcce4d19985c1089eaf77901901