Install docker registry cert at bootstrap
Docker registry certificate is first generated when the functionality
is started in bootstrap using a self-signed CA. Later on, we are
creating the certificate using a issuer customized ('system-local-ca').

Cert-mon is able to pick up the new certificate and install it after
the first unlock, but there's a window of time were the system is still
trying to use the old certificate. This change installs the new
certificate directly after the creation, at the end of bootstrap,
making the system use the correct certificate before Cert-mon kicks in.

Test plan:
PASS: Deploy a system with the feature flag enabled in the localhost
      file ('create_platform_certificates'). Verify that the system
      bootstraps correctly and the pods from default applications are
      able to pull the required images from the local registry.

Story: 2009811
Task: 49123

Change-Id: I0000ad23a36f512d4a33fd34961f544fd3a0dbe9
Signed-off-by: Marcelo Loebens <Marcelo.DeCastroLoebens@windriver.com>