Fix remote copy during rehoming
Fix error introduced by a change in the role that performs
system-local-ca RCA install:
https://review.opendev.org/c/starlingx/ansible-playbooks/+/900637

The issue was caused by not enabling the parameter remote_src in a
usage of ansible copy where the src and target files should be in the
subcloud (default is false, ansible search for the source file in the
managing host, i.e. the SystemController).

Test plan:
PASS: Deploy one DC w/ SystemController + one SX subcloud.
      Deploy other SystemController.
      Rehome the subcloud from one SystemController to the other.

Story: 2009811
Task: 49122

Change-Id: I517142ed1f2ad4f9a667014958fb14d8a5f0d137
Signed-off-by: Marcelo Loebens <Marcelo.DeCastroLoebens@windriver.com>