Enforce password change when installing via rvmc
When installing a subcloud via rvmc, it's possible that the BMC
reports a successful operation when booting with new install media
without actually installing the subcloud. To prevent this, a new
variable called enforce_password_change was created, failing the
installation if there is an error during initial password change
(observed if the system wasn't correctly reinstalled).

Test plan:
  - PASS: Perform a subcloud add with install forcing a failure by
          changing the password before the playbook and verify the
          installation fails as expected.
  - PASS: Perform a normal subcloud add with install and verify
          the operation completed successfully.
  - PASS: Deploy a standalone SX and verify the bootstrap completed
          successfully.
  - PASS: Upgrade a subcloud from stx6 to stx8 and verify the
          operation completed successfully.

Closes-Bug: 2044564

Change-Id: I80bee246dedfdf9688507c3529d7d080992da08b
Signed-off-by: Victor Romano <victor.gluzromano@windriver.com>