Fix wipe_osds.sh script
An inconsistency was found in the wipe_osds.sh script, where it
reversed the ceph data and ceph journal partitions. This happened
because the partition number is obtained through a counter, which
is incremented with each for loop. Since the lsblk command used to
obtain the partitions, the /dev/sdX2 partition was returned as the
first in the list. This way, the script obtained information from
one partition while it was processing another.

To resolve this, sfdisk was used to obtain disk information and
udevadm was used to know exactly which partition number is being
processed. These commands were used based on the sysinv's
partition_info.sh script.

Finally, some other points in the script were improved according
to analysis by the shellcheck tool.

Test Plan:
  - PASS: (AIO-SX) Replace wipe_osds.sh with changes,
          run and check script output
  - PASS: (AIO-SX with multipath) Replace wipe_osds.sh
          with changes, run and check script output
  - PASS: (AIO-SX) Optimized restore with wipe ceph
  - PASS: (AIO-DX) Legacy restore with wipe ceph
  - PASS: (AIO-DX) Upgrade with Ceph storage backend

Closes-Bug: 2044806

Change-Id: If1629e53f9a298d8202fb881bd13b6772da9ee38
Signed-off-by: Erickson Silva de Oliveira <Erickson.SilvadeOliveira@windriver.com>