Enable support for PVC snapshots by default during bootstrap
Previously, the “enable_volume_snapshot_support” variable was false
by default. As a result, when running the bootstrap playbook, the
CRDs, and the snapshot-controller were not created to support taking
snapshots of PVCs.

With the change, when running the bootstrap playbook, the CRDs and
the snapshot-controller will be created by default according to the
K8s version on the system.

The main benefit of this change is that CSI apps will not need to
worry about creating CRDs and the snapshot-controller to be able to
take snapshots of PVCs, in addition to avoiding possible problems,
since the resource creation process and the version used will be in
common for all CSI apps.
Furthermore, if the user is not going to use these resources, he can
choose to run the bootstrap manually by passing the variable to
false as a parameter to not create.

Currently, the version of the CRDs and snapshot-controller present
in the system is very old, referring to k8s 1.19 with the
'snapshot-controller:v2.0.0-rc2' image. Due to this, the creation of
VolumeSnapshots does not occur correctly with new versions of some
CSI apps.
Therefore, the current change depends on the review below, where
the CRDs and snapshot-controller are updated according to the
K8s version, working as expected.

* This is a new review after the revert of:
https://review.opendev.org/c/starlingx/ansible-playbooks/+/902811

Test Plan:
 PASS: Build an image with the code change
 PASS: AIO-SX fresh install + Check if the CRDs
       and snapshot-controller are created
 PASS: AIO-DX fresh install + Check if the CRDs
       and snapshot-controller are created
 PASS: Standard fresh install + Check if the CRDs
       and snapshot-controller are created
 PASS: Storage fresh install + Check if the CRDs
       and snapshot-controller are created

Closes-bug: 2045897

Depends-On: https://review.opendev.org/c/starlingx/ansible-playbooks/+/904359

Change-Id: Ifa8e9fc107a264c7df7eb8e0d1f8fa71e6fb7598
Signed-off-by: Gabriel de Araújo Cabral <gabriel.cabral@windriver.com>
Signed-off-by: Erickson Silva de Oliveira <Erickson.SilvadeOliveira@windriver.com>