Intergrate cert recovery into rehome playbook
This commit integrates the recover-subcloud-certificate role into the
rehome_subcloud playbook, enables recovering a subcloud's certificates
if they are expired in the subcloud rehoming process.

Test plan:
1. Passed - rehome an AIOSX subcloud w/o cert expired.
2. Passed - rehome an AIODX+ subcloud w/o cert expired.
3. Passed - deploy an AIOSX subcloud as 11 years ago, correct its time
and restart it. Rehome the subcloud. The 1st attempt fails with log to
recover the certs(docker registry cert) which request manual recovery.
Manually recovered the certificate, then delete the subcloud and
rehome it again. Verify the rehome is successful and the subcloud can
be online and insync after management.
4. Passed - repeat the steps in TC3 on an AIODX+ subcloud, verify the
subcloud can be online and insync after the 2nd attempt of subcloud
rehoming.

Story: 2010815
Task: 49265
Signed-off-by: Yuxing Jiang <Yuxing.Jiang@windriver.com>
Change-Id: I902014aed758a894d4ed9ed83394424484817337