Fix system_platform_certificate upgrade issue
The 'system_platform_certificate' override should not be
provided during upgrade. A fail task inserted to check if the
variable is defined is being triggered during upgrades.

This is happening because the playbook is importing the default
values for this variable defined in specific task at the beginning of
the play, defining the variable and triggering the fail task.

This review removes the default definition for
system_platform_certificate, since the certificates template creation task, that uses the variable values, already has defaults for most of
the required fields defined in it's own default variables. I'm
including the only missing value ('renewBefore') in said template
task defaults.

Also, set the 'true' flag in the jinja default() filter in the template
file at every usage regarding the generated platform certificates, since
none of them is supposed to accept a empty value.

Test Plan:
- Deploy AIO-SX
- Execute platform certificates update playbook.
- Upgrade AIO-SX (passed the stage where the error happens, but
  upgrade isn't successful due to other blocking issues in the
  playbook).

Story: 2009811
Task: 49378

Change-Id: I9fd9871f4c41106a224906d9a60987885d2d8fce
Signed-off-by: Marcelo Loebens <Marcelo.DeCastroLoebens@windriver.com>