Fix for snapshot-controller image pull error
After rebooting the active controller, the snapshot-controller pod
is no longer able to pull the image, changing to "ErrImagePull"
status.

To fix the error, the 'imagePullSecrets' field was added to the
service account of the snapshot controller from K8s 1.25 to 1.27,
specifying the secret required to authenticate to the registry
and pull the image.

Test plan:
 PASS: AIO-DX fresh install
 PASS: Reboot the active controller, checking when it returns that
       the snapshot-controller pod is running
 PASS: Reboot both controllers, checking when they return that the
       snapshot-controller pod is running

Closes-bug: 2051844

Change-Id: I68f6014898e193e561774f298c7997a3ce89a4eb
Signed-off-by: Gabriel de Araújo Cabral <gabriel.cabral@windriver.com>