Fix incorrect system images for K8s 1.25
In the review [1], a new system-images.yml for K8s 1.25 was created,
replacing a symbolic link that pointed to K8s 1.26.
The new system-images.yaml for K8s 1.25 was a copy from K8s 1.26,
with the only difference being the version of the snapshot-controller
image.

[1] https://review.opendev.org/c/starlingx/ansible-playbooks/+/887797

The issue with the K8s 1.25 images arose because, while the review
was open, some images were updated in the K8s 1.26 system-images.yml,
which, in the master branch, they were also used for K8s 1.25.
However, these changes to the images were not reflected in
the system-images.yaml of K8s 1.25 created in the review that was
mentioned. Consequently, when merged, K8s 1.25 ended up containing
the outdated images.

The current commit updates the images that were impacted on K8s
1.25.

Updates missed due to new system-images.yml for K8s 1.25:

https://review.opendev.org/c/starlingx/ansible-playbooks/+/890987
https://review.opendev.org/c/starlingx/ansible-playbooks/+/898041
https://review.opendev.org/c/starlingx/ansible-playbooks/+/899271
https://review.opendev.org/c/starlingx/ansible-playbooks/+/899982

Removing the RVMC image as the RVMC script was decontainerized in
the review:

https://review.opendev.org/c/starlingx/ansible-playbooks/+/897490

Test Plan:
 PASS: Run ansible bootstrap playbook with new images on
       k8s 1.25.3 (AIO-SX fresh install with K8s 1.25)
 PASS: Verify all images are downloaded and tagged with
       local registry after installation
 PASS: Check that all pods are healthy and using the
       correct image after installation

Closes-bug: 2052728

Change-Id: I269bef9fd679061797e5042263f59cabc2a6e822
Signed-off-by: Gabriel de Araújo Cabral <gabriel.cabral@windriver.com>