Add wildcards argument in recover-ceph-data
The restore-ceph-data playbook extracts a file using pattern
matching characters when the backup was generate on
controller-1 (AIO-DX). On centOS the command work as it is,
but on Debian is necessary to use the --wildcards argument
to enable the pattern matching, or it will raise an exception.

To resolve this issue, the argument was added to the command.

Test Plan:
PASS: (AIO-DX) Set controller-1 as the active controller
PASS: (AIO-DX) Perform backup and restore

Closes-bug: 2052968

Change-Id: Ie373e8d2bceb8f648ed68c62eefc8867d74a8588
Signed-off-by: Erickson Silva de Oliveira <Erickson.SilvadeOliveira@windriver.com>