Optimize ip address validation
This commit updates address validation script to accept a list instead
of individual item. This will reduce processing time by 2/3.

TEST PLAN:
1. PASS: Deploy a controller and verify the ip validation execution
time in ansible.log. Re-deploy the controller with the changes in
the bootstrap playbooks. Verify that the task took less time to
finish.

2. PASS: Deploy a subcloud in a DC environment and verify the ip
validation execution time. Deploy an another subcloud with the
playbook changes and verify that the ip validation task took less
time to finish.

3. PASS: Deploy a simplex system controller and change two docker
registries to invalids addresses. Run bootstrap Verify that the
playbook fails with a list of the invalid addresses.

Story: 2011035
Task: 49569

Signed-off-by: Gustavo Pereira <gustavo.lyrapereira@windriver.com>
Change-Id: I2b32c71116a0602d386bc1849fd80eaa081dd939