Skip cert file transfer and validation during restore
In restore, the certificates should only be restored from the backup.

This commit skips task to tranfer and validate the certs during
restore to prevent the failure of validation during subcloud restore.

Test plan:
1. Passed - verify bootstrap a DC(DX system controllers and a DX
subcloud).
2. Passed - verify backup and restore the DX subcloud. The restore
playbook passed bootstrap/validate-config role, but was not able to
reach the bringup-kubemaster role due to:
https://bugs.launchpad.net/starlingx/+bug/2056054.

Closes-bug: 2054165

Signed-off-by: Yuxing Jiang <Yuxing.Jiang@windriver.com>
Change-Id: I9cdb1b1ea28b35a407b05525ccfa5a25bd3b3004