Rehome: Increase timeout for certs to be installed
This commit addresses 2 related bugs, in the sense that they
are related to ansible tasks that wait for certificates
to be installed.

Task 'Check admin-ep-cert.pem updated' depends on cert-mon to
install the certificate. Cert-mon may fail and reattempt after
10 minutes. This change increases that timeout for this task
to be larger than that. I'm also decreasing the delay in a
half as it is a quick stat operation, to allow it to be
detected quickly in most cases where the first cert-mon attempt
works.

Task 'Verify if there are 250.001 (config out-of-date) alarms'
is dependent on puppet to apply a config change to install
the certificate and sysinv to clear the alarm. When I
reproduced this issue it took 1 minute longer for the alarm
to clear. This change increases the timeout of the task in
about 50%.

Test Plan:

PASS: Rehome a subcloud with 1200 ms latency injected and
      50% cap on CPU capacity.

Closes-Bug: 2054462
Closes-Bug: 2054463
Change-Id: I017fab0ccb13629c63a7cd855470f0a777f06e22
Signed-off-by: Rei Oliveira <Reinildes.JoseMateusOliveira@windriver.com>