Skip N3000 reset during AIO-SX upgrade migration
When N3000 is used by an app, an N3000 reset will be triggered
during upgrade.  If the system is AIO-SX, the migration will fail
because a certain image is required, because image are not available
at this point in the upgrade.

The N3000 reset can be skipped by created a volatile flag.
This flag is cleared on reboot so the N3000 reset will be
executed after unlock.

TEST PLAN
* AIO-SX optimized upgrade, stx6 to stx8, with FEC app
* AIO-SX optimized upgrade, stx6 to stx8, without FEC app

Closes-Bug: 2055049
Change-Id: Ia026cc6ed05bd3183555b73eeadc84e3ab180ffa
Signed-off-by: Joshua Kraitberg <joshua.kraitberg@windriver.com>