Replace sftp server bin with debian's path
The aio-sx restore playbook will restore the /etc/sshd dir
from the centOS backup when upgrading from centos to debian.
This will result in etc/ssh/sshd_config pointing to a sftpserver
path that does not exist in debian. This is reproducible
on AIO-SX standalone and subclouds, but not AIO-DX, since
SX and DX have different upgrade paths.

/usr/libexec/openssh/sftp-server in centos.
/usr/lib/openssh/sftp-server in debian.

This is a simple change that simply replaces the wrong path
with the valid path for debian.

Test case:

PASS: Run AIO-SX upgrade then verify that 'sftp controller-0'
      goes thru and the connection is established.
      'Get a file' and verify the download is successful.

Closes-Bug: 2055324
Change-Id: I15a47dcb0ea3dae49148b4b38f722923409b41ed
Signed-off-by: Rei Oliveira <Reinildes.JoseMateusOliveira@windriver.com>