Avoid hard-coded wait for kube-system pods startup
Replaced the hard-coded 120-second pause for kube-system pods startup
with a periodic conditional (retry-until-delay) pause. Overall, this update reduces the bootstrap time by ~120 seconds.

In most cases, the pods are created significantly before the timeout
period, even before the initiation of the pause task. This suggests that while
the pause is necessary for rare circumstances, it was used unnecessarily
at most times. That is, the previous fixed wait time was suboptimal as it prolonged the bootstrap process for the typical case.

Test Plan:
1. PASS: Verify AIO-DX install/bootstrap (virtual and physical lab)
2. PASS: Verify full DC system (System Controller + 3 Subclouds)
         install/bootstrap (virtual lab)
3. PASS: Run multiple iterations of install to
         ensure no intermittent issues introduced for bootstrap.
4. PASS: Observe kube-system pod objects (lables, pod names, etc)
         during pre/post the wait task, and after
         the system is fully bootstrapped. Ensure new task
         is implemented correctly based on pod objects.
5. PASS: Run the new task commands in isolation and verify
         correct behavior:
          - Grep string pattern generated as expected:
            7 search strings ORed together
          - kubectl command output:
            kube-system pods labels listed
          - "wc -l" returns expected number

Story: 2011035
Task: 49662

Change-Id: I16ef2d7718efa9dc5245f2877eacb4f8515419e8
Signed-off-by: Salman Rana <salman.rana@windriver.com>