Increase timeout for cert config to apply
This commit addresses a bug where a task that waits for
certificates to be installed and the puppet config
to be applied, times out and fails.

Task 'Wait for .config_applied file stat to change'
currently only waits about 3 minutes. In high stress
scenarios like parallel rehoming, this is not enough. For
instance, one scenario this task failed, it took additional
4 min for the alarm to clear and config to be applied.
This fixes that by increasing the number of retries.

I'm also decreasing the delay in half as it is a quick stat
operation, to allow it to be detected quickly in most cases
and results in quicker detection overall.

Test Plan:

PASS: Rehome a subcloud with 1200 ms latency injected and
      50% cap on CPU capacity.

Closes-Bug: 2056183
Signed-off-by: Rei Oliveira <Reinildes.JoseMateusOliveira@windriver.com>
Change-Id: Iad009087c3bbf0f27da93912b8096392034be743