Only wait for essential pods in cert recovery
The certificate recovery role will trigger a restart of every pod
in the k8s cluster so that they can be updated with the latest
certificate information.

After pods restart the procedure waits every pod to recover and become
READY. This change modifies that behaviour to only wait for essential
pods to recover, being those in the core namespaces armada,
cert-manager, flux-helm and kube-system.

Test case:

PASS: Run certificate recovery with crashing pods in a custom namespace

Closes-Bug: 2058751

Signed-off-by: Rei Oliveira <Reinildes.JoseMateusOliveira@windriver.com>
Change-Id: I3ea403a3e324ecbb5f2c1f56d6ce1c8bd80fabee