Increase timeout to unlock host on cert recovery
The certificate recovery role may try to unlock a host while it's
on a reboot loop. In order for the host-unlock command to be accepted
to host needs to be online.

For slower hardware the current 2 min timeout is not enough.
This commit changes it to 5 min.

Test case:

PASS: Run certificate recovery rehoming with multi node systems.

Closes-Bug: 2058426

Signed-off-by: Rei Oliveira <Reinildes.JoseMateusOliveira@windriver.com>
Change-Id: I7a7faf7707a19f477a968766cbe383e7dfdbd1cd