Remove conditional statement for enabling IPv6 firewall in Calico
This change removes the conditional statement around the section that
enables the firewall for IPv6 in Calico. The IPv6 firewall will be
permanently enabled regardless of the setup, so that even if IPv6 is
unused, traffic will be blocked.

Test plan
=========

The tests for https://review.opendev.org/c/starlingx/config/+/915508
also cover this change.

Story: 2011027
Task: 49816
Depends-On: https://review.opendev.org/c/starlingx/stx-puppet/+/915509
Change-Id: I986f361493b29596851e781632c782c92ec22546
Signed-off-by: Lucas Ratusznei Fonseca <lucas.ratuszneifonseca@windriver.com>