Create barbican secrets before sysinv db update
This commit creates the Barbican user, service, and
endpoints before the 'update sysinv db' task. This
enables the creation of the registry secrets to be
used to create service parameters for the registries.

Test Plan:

PASS: Perform a complete deploy in a DC environment.
PASS: Check that all service parameters for Docker are created.
PASS: Successfully apply the platform-integ-apps.

Closes-Bug: 2065317

Change-Id: I259e176a4a6309ca8748aef37e137e0c6e0894b9
Signed-off-by: Hugo Brito <hugo.brito@windriver.com>