hc_vault: do not become root for localhost
When performing remote ansible execution the playbook tries to become
root when delegated to the localhost (inspection_target for remote
execution). The error occurs at task "Validate initial_backup_dir
exists".

Reserve "become" for when it is needed, at task "Create vault
subdirectory in initial_backup_dir", and transfer of the tarball when
not on_box_data.

Also move the creation of the subdirectory to follow after the tests for
readiness of the vault app.

Implement the logic as well for role vault/vault_restore.

Test Plan:

PASS  local and remote playbook execution

Story: 2011073
Task: 49842

Change-Id: I241d3c7b2594e2ef308c12dcf5128ebe079a7f24
Signed-off-by: Michel Thebeau <michel.thebeau@windriver.com>