Check subcloud connectivity during CA update
This review will be creating a new role to common, that is
check-connectivity, that checks if a subcloud is reachable.
The code was moved from common/prepare-env to a new path
common/check-connectivity, and is also imported by
common/prepare-env.

This review will be also adding a role to the tasks in "Gets a summary
of certificates on system", this role, common/check-connectivity,
will check the connectivity of the subcloud before the desired task is
executed, preventing the task to execute when the subcloud is offline.

Test Plan:
PASS: Run 'migrate-certificates' playbook with online subcloud and
      verify if it finish successfully.
PASS: Run 'migrate-certificate' playbook with offline subcloud and
      verify if the playbook finish with connection error.
PASS: Run 'migrate-certificate' playbook with online subcloud in
      beginning of the run and then turn off the subcloud to check
      if the playbook will return the connection error.
PASS: Run backup and restore and verify if the playbook run
      common/prepare-env with no error and finishes correctly.

Closes-Bug: 2067619

Change-Id: Iad85a8769e24b4cc72dace73e0ad7f25b9742983
Signed-off-by: Karla Felix <karla.karolinenogueirafelix@windriver.com>