Enable certificate tasks for enrollment
This commit enables the certificate tasks for the enrollment playbook.
- rehome-subcloud/update-ca-cert
- common/update-sc-cert

We install the file admin-ep-cert.pem through ansible because cert-mon
is running in a "regular system" mode as opposed to DC mode at the
moment, and does not monitor the admin-ep-cert or install a new
copy.

Test Cases:

The test environment is a standard system controller with a single
node subcloud, installed with the existing factory install procedure.

PASS: get certificate on the subcloud, ensure sc-adminep-certificate
      is updated
PASS: ensure /etc/ssl/private/admin-ep-cert.pem contains the cert
      and key, like a regular DC deployment

Story: 2011100
Task: 50226

Change-Id: I12b5fb8a7b13fdcec6985a599347ff184e6287e0
Signed-off-by: Jerry Sun <jerry.sun@windriver.com>