Add label component=platform to pods
Add the label "app.starlingx.io/component=platform" to pods
to ensure kubernetes run it on the platform cores.

Additionally, the review [1] introduced a problem when
changing the chdir path in the trident install task.
This resulted in the installer being unable to locate
the custom templates.

[1]: https://review.opendev.org/c/starlingx/ansible-playbooks/+/887424/13/playbookconfig/src/playbooks/roles/k8s-storage-backends/netapp/tasks/main.yml#86

Test Plan:
PASS: Install trident
PASS: Check if pods have the label
'app.starlingx.io/component' with the value 'platform'

Story: 2010612
Task: 50258

Change-Id: I051c328edddf5b997c192e96f9be4ed34743b554
Signed-off-by: Erickson Silva de Oliveira <Erickson.SilvadeOliveira@windriver.com>