Improves kubelet configuration for dual-stack node addresses
This change ensures the node-ip flag in the kubelet is set correctly
for single-stack and dual-stack deployments. Additionally, it
validates that all cluster network CIDRs (pod, service, and host) are
configured for dual-stack functionality when the feature is enabled.

Test Plan
=========

All tests checked, after bootstrap, the configuration in:
- configmap kube-proxy
- configmap multus-cni-config.v1
- configmap kubeadm-config
- pod kube-controller-manager-controller-0/1
- pod kube-apiserver-controller-0/1
- daemonsets.apps calico-node
And started new pods and services on both stacks to check traffic
intra and inter nodes

All installations below were done with single-stack (IPv4 and IPv6)
and dual-stack:
- IPv6 primary, IPv4 secondary
- IPv4 primary, IPv6 secondary

[PASS] Install AIO-SX
[PASS] Install AIO-DX
[PASS] Install Standard

Story: 2011027
Task: 50319


Change-Id: Ic59008bb06b6697cfc1e1d0ba353ebedecd2481b
Signed-off-by: Andre Kantek <andrefernandozanella.kantek@windriver.com>